cdk/clientell/requirements.txt
aws-cdk.aws-s3
aws-cdk.core
aws-cdk.aws-ec2
aws-cdk.aws-kms
aws-cdk.aws-codebuild
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline-actions
aws-cdk.aws-ssm
aws-cdk.aws-codedeploy